+++ plugin/vscode/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/vscode-extension-v1.0.1...HEAD)
 ------------------------------------------------------------------------
+### Added
+- Drop stale results when a file changes outside the editor
 
 [1.0.1](https://github.com/jshwi/docsig/releases/tag/vscode-extension-v1.0.1) - 2026-07-10
 ------------------------------------------------------------------------